pyproject.toml
[project]
name = "assignment-4"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "openai-agents>=0.2.11",
    "python-decouple>=3.8",
]
